+++ CPE200Lab1/CPE200Lab1/CPE200Lab1.csproj
@@ -45,6 +45,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="basicCalculatorEngine.cs" />
     <Compile Include="CalculatorEngine.cs" />
     <Compile Include="ExtendForm.cs">
       <SubType>Form</SubType>
